requirements.txt
certifi==2019.3.9
chardet==3.0.4
disco-py==0.0.12
gevent==1.2.2
greenlet==0.4.15
holster==2.0.0
idna==2.6
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.44.0
